+++ src/main.rs
@@ -1,3 +1,5 @@
+mod ui;
+
 use std::io;
 
 use crossterm::event::{self, Event, KeyCode, KeyEvent, KeyEventKind};
@@ -43,37 +45,9 @@ pub enum CurrentNode {
 
 impl App {
     pub fn run(&mut self, terminal: &mut DefaultTerminal) -> io::Result<()> {
-        let tasks = ["A", "B", "C", "D"];
         while !self.exit {
             terminal.draw(|frame| {
-                let area = frame.area();
-                let window_chunks = Layout::default()
-                    .direction(ratatui::layout::Direction::Horizontal)
-                    .constraints([Constraint::Percentage(50), Constraint::Percentage(50)])
-                    .split(area);
-
-                // Draw tasks
-                let tasks: Vec<ListItem> = tasks
-                    .iter()
-                    .map(|i| ListItem::new(vec![text::Line::from(*i)]))
-                    .collect();
-                let tasks = List::new(tasks)
-                    .block(Block::bordered().title("List"))
-                    .highlight_style(Style::default().add_modifier(Modifier::BOLD))
-                    .highlight_symbol("> ");
-
-                let mut block_left = Block::bordered().title_top("Left").border_type(ratatui::widgets::BorderType::Rounded);
-                let mut block_right = Block::bordered().title_top("Right").border_type(ratatui::widgets::BorderType::Rounded);
-
-                let active_style = Style::default().fg(ratatui::style::Color::Blue);
-                match self.current_node {
-                    CurrentNode::Left => block_left = block_left.border_style(active_style),
-                    CurrentNode::Right => block_right = block_right.border_style(active_style),
-                };
-
-                frame.render_widget(&block_left, window_chunks[0]);
-                frame.render_widget(&tasks, window_chunks[0]);
-                frame.render_widget(&block_right, window_chunks[1]);
+                ui::ui(frame);
             })?;
             self.handle_events()?;
         }
